refactor(conversation_tab): create a prompt_attachment component for prompt text and attachment list
